REBSER vs REBVAL vs UTF-8 "pointer detection"
This patches a design from Ren-C into the code, using %sys-base.h
and %sys-flags.h (which I've relicensed as Apache to make shareable.)

You can give Detect_Rebol_Pointer() an arbitrary pointer and it will
come back with:

    DETECTED_AS_UTF8
    DETECTED_AS_REBVAL
    DETECTED_AS_REBSER
    DETECTED_AS_END
    DETECTED_AS_FREE
    DETECTED_AS_WILD

It's tried and tested code, which has been carefully crafted in order
to not run afoul of strict aliasing problems in C.  This has been
working for many years on every platform tested for.

This helps some optimizations, because you can have a slot that holds
a pointer and know what it is without needing to separately store a
flag or information to say what kind of pointer you store.

Other applications are e.g. debugging routines (being able to take
a pointer and figure out if it's a pointer to a REBSER or REBVAL
automatically, so you can have a single Probe() function vs. needing
Probe_Rebser() or Probe_Rebval(...).

(Some extra work would be required making free poolunits use the
FREE_POOLUNIT_BYTE, in order to distinguish empty UTF-8 strings from
a freed unit--which can be nice for debugging.)

Note that critically, the ability to "sniff" a pointer and know if
it is UTF-8 or a REBVAL allows the support of variadic APIs:

  https://rebol.metaeducation.com/t/no-preprocessing-no-ffi/2224